Fixed Linux compatibility issues with test_features. Converted samples to Linux.
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@112 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/tests/test_feature/source/testexecute.cpp
@@ -36,9 +36,9 @@ bool TestExecute()
 
 	
 	asIScriptContext *ctx = engine->CreateContext();
-	assert(ctx->SetUserData((void*)0xDEADF00D) == 0);
-	assert(ctx->GetUserData() == (void*)0xDEADF00D);
-	assert(ctx->SetUserData(0) == (void*)0xDEADF00D);
+	assert(ctx->SetUserData((void*)(size_t)0xDEADF00D) == 0);
+	assert(ctx->GetUserData() == (void*)(size_t)0xDEADF00D);
+	assert(ctx->SetUserData(0) == (void*)(size_t)0xDEADF00D);
 	ctx->Release();
 
 	engine->Release();
